+++ src/pages/query.tsx
@@ -1,3 +1,0 @@
-import Query from "lib/pages/query";
-
-export default Query;